Switch to C++ for builtin modules.
The main reason for this move is that implementing advanced vector
handling for derivatives will be easier in C++. This can be revisited
later.

One big pitfall with C++ is the name mangling: functions get renamed by
the compiler, so that lookup by name does not easily work any more.
To avoid the issue, functions that are externally visible should have
an 'annotate' attribute with the "real" name. This gets converted into
an LLVM function attribute after parsing the IR, which then serves to
find necessary functions.