Removed the unfinished image writing mode of the ForceViz modifier.
The code was intended for baking texture info into an image texture.
This is quite involved and should better use a texture in combination
with existing bake operators. Important is to not allow this texture in
rendering, because oversampling can lead to a large and slow number of
evaluations (baking otoh should need no more than a single eval per
fragment).