Fix for registering RNA classes with registered parent functions.
The recursive registration for functions uses a simple int array for
flagging found functions. This is shared with the parent, so any child
functions will overwrite parent function flags for the same index
(this is not overloading).

The patch now requires using a larger array that can accomodate all
registerable functions from parents and children. It puts flags for
parents first, then children.