Improved fluid sources implementation.
Generating points based simply on a fixed density or density-rate
only works when the fluid has space to move into, otherwise the point
density is simply increased without adding more fluid (since points do
not actually represent mass).

The new method aims to use grids to estimate density and only add as many
points per cell as are needed for keeping a sufficient level of detail.
This can be used to either create reservoirs that are always filled with
liquid, or to create water inflows with a certain pressure by using
a starting velocity.