Split the "uber-eval" function for object data into individual operations based on object type.
This means making a dedicated data generation function for object types that need it, and a few utility functions for particles and extra driver evaluation (hopefully soon to be deprecated).

Remaining "uber" functions in the depsgraph should be split in a similar manner in future.

The splitting allows use to go deeper into the design of certain object types (meshes in particular) and improve these without having to deal with all the other types.