Reference counting for mesh_ptr variables on the stack.
Any extensive data stored as a pointer on the stack should be ref-counted,
so that it stays alive only as long as used by some node.

The way this works is by enclosing instructions that access the stack
entry with 'init'/'release' opcodes. These are added automatically based
on output socket types (nodes don't have to be created explicitly for them).