Extension of the new dependency graph relations to support object nodes.
This generally triggers a mesh data recalc when *anything* inside the
object node tree is changed. Eventually it should be more specific to
changes in a geometry node tree, but the current design of deps nodes
does not reflect such internal node semantics nicely (although new
depsgraph will allow us to do it properly in the future).