Horrible fix for depsgraph/notifier updates through property changes in geometry nodes.
This is really awful, but we don't have a good generic way of defining node tree relations
in the dependency graph yet. More importantly, the notifier system is not up to the task -
putting the responsibility for notifying into operators requires anticipating their effects,
which is not good at all for node operators which affect geometry through complex processing.
The depsgraph really should be handling this.