New node "Curve Path" for reading interpolated curve data.
This introduces an improved way of handling ID relations inside the
node tree. Instead of storing simple indices to an object array,
the nodes now generate a 'key' (currently just ID name + lib name).
At eval time this key is then looked up from a hash table in globals,
to produce the actual pointer to the object.

Currently the node will not work, because the curve is not yet registered
in the globals map. This should happen in conjunction with depsgraph
updates.