Minimal RNA API of the depsgraph for registering object node dependencies.
This simply reflects the existing C API for the depsgraph, as already used
by modifiers. The object nodes gain two callbacks, one for depsgraph
relations and one for registering globals. Since both of these should
require the same objects and other ID blocks to be registered, they share
a common `relations_update` function on the pynode level for convenience.