Use the relative transform of target objects and the modified object to adjust modifiers.
This affects modifiers which use an external target object (currently
boolean and curve path). The modified object is a graph input argument,
which the target object is looked up via its key. The relative transform
means the values calculated from external geometry are always in the
modified object's local space.

Note that transform and its inverse are provided explicitly to nodes.
This will allow optimization for constants later on.