Added a simple operator and panel for one-click debug display of bvm node visualizations.
This was formerly part of an addon, but is useful enough to include directly in the branch
for debugging purposes. It uses the existing graphviz dump functions to write into a
temp dot file, then runs graphviz on it and opens the resulting image in the standard
system image viewer.

This is tailored for linux systems, but that shouldn't be a big issue since most developers
(who are the target audience) use this platform anyway.