Split off from the bvm::Function class the basic instruction list functionality.
The point of this is to keep type definitions out of the core bvm eval code.
TypeDesc is only used for the outermost layer of bvm::Function in order to
copy arguments to and results from the eval stack. OpCodes themselves are
agnostic to higher type definitions and just work with raw data.