Fix node socket update callbacks exiting early because of missing output connection.
Outputs are currently identified in the nodes via the DO_OUTPUT flag. This is not
nicely accessible for pynodes and needs to be reimplemented in a proper way.
Until then this optimization for node recalculation should just be ignored for texnodes,
so nodes now have a callback for checking usage, and will always be considered used by default.