Fix for depsgraph building of modifiers, missing nodes for ID blocks (textures).
The updateDepsgraph callback for modifiers needs to be used in both passes of
depsgraph building, once for creating nodes and then for relations. Without the
initial node building pass some datablocks like textures can be missing, if they
are not incidentally built by other means such as a render texture.