Another depsgraph fix, internal texture relations need to be built explicitly.
Like node trees, textures are only optionally constructed in the depsgraph when something
uses them. When a modifier (or other potential future user of the "handle" system)
declares a texture dependency, the internal relations of that texture are not necessarily built -
this only happens when the texture is part of a material texture stack. So we need to also
construct the internal texture relations (lazy init) in addition to the external modifier->texture
dependency.