Use the Strands modifier as a simple container for strand control curves.
These curves are NOT intended for final rendering! They are abstract
editable data (control strands) from which the "actual" visible strands
have to be generated in some way.

This is largely equivalent to the parent/child dichotomy in old particles,
but enforces the separation right from the start (control curves are never
rendered).