Revert "Remove redundant dependency declaration from pynodes, which creates a circular dependency."
This reverts commit ca85cb436db7dc9ff5e8df3172987b09f5a071c4.

Reverting the commit will break depsgraph updates for regular bvm functions again, but is necessary
for making shader updates working. The problem here is that the update operations for _dependent_
data (bvm functions, shaders, etc.) should not actually be part of the node tree, but of the
respective dependent data components! So the previous method of invalidating shaders as part
of the node tree operations is incorrect and the operation should be moved to some other place.
Perhaps we need a dedicated depsnode for the bvm function cache, since there is no single "owner"
of these functions that could take care of invalidation in the depsgraph.