Use a "pass" concept for strand shader organization, like the GPUMaterial system.
This will allow easier addition of other shader features later on. The system
should be merged with the GPUMaterial code, but currently the material system
is too unwieldy to integrate the strand displacement code easily. With the 2.8
viewport rewrite all of this will have to be rewritten as well anyway.