Add a cache structure for collision contact points.
This should serve as a universal interface between possible collision
detection systems (BVH, Bullet) and the various collision response systems.
Using an new data structure means that the two sides of a collision
handling feature (detection and response) can be treated independently,
reducing code dependencies and complexity.

The properties of contact points are modeled after btManifoldPoint from
Bullet and may need to be adjusted in the future.