Simple contact detection method for hair editing, using the typical BVH closest point method.
This method is just a placeholder, in the future it would be worthwhile to use Bullet collision
detection with ghost objects for this.

Also the code should eventually be abstracted a bit to make it agnostic to mesh formats, so it
can be used for BMesh as well as DNA strand data or physics simulation.