Fix for frequent unnecessary shader recompilation during hair editing.
The depsgraph component node for invalidating the strand shader was located in
the GEOMETRY component of the object. This component by default gets tagged along
with every other object part during editing. Only standardized way of avoiding this
would be putting it in a different datablock, but the shader is very much tied to
the particular object due to the displacement function from its node tree.

For now the solution is a new component GEOMETRY_SHADER, which does not get tagged
when the ID node is tagged. It still gets regular updates from node tree changes.
This seems to be the least invasive way of avoiding unnecessary performance drops,
and can be removed and replaced easily later on without disrupting other code work.